Perform GC_STRESS in alloc() instead of eval0().
To catch as many dangling pointers as possible.  It's not a
mathematical proof if it passes.  Some can escape since objects in
question might not be moving and check_lispptr() could, for reasons
yet unknown, give false positives.  check_lispptr() will have to go
all over the heap to make sure that a pointer refers to the first
byte of a valid object.  That is rather time-consuming but should
be doable on modern targets (currently TARGET_UNIX).  Also, all
objects should move to ensure that pointers, which haven't been
placed on the GC stack to get relocated, are spotted by
check_lispptr().  Last but not least check_lispptr() should be
called whenever a pointer is moved to or from places.

These are things that are easy to implement but going down easy
road should uncover just enough bugs to keep one busy for a couple
of days.