Add check_lispptr_addr().
Walks all of the current heap and checks, if an address matches the
start of an object.  If not, an internal error is issued.

Fails in the GC's sweep phase already, although the pointer does
point to the start of an object and 'heap_end' is beyond.  The
object pointed to is unused though.  That does not matter to
check_lispptr_addr().