docs(compiler): Expression expansion is not part of the first version.
It is required for transpiling to other languages by assigning new
variables in this version.  If the stack was used to pass arguments,
like in the second version of the compiler which also compiles
function calls, it'd even be simpler, not creating assignments at all,
but pushes onto the object stack, if that's not done in function
epilogues.  Depends on the target which scheme is favourable.